lowlight/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
tqdm>=4.66.0
matplotlib>=3.7.0
tensorboard>=2.13.0
scipy>=1.11.0
lpips>=0.1.4
pandas>=2.0.0

